keyword_analyzer/requirements.txt
lxml==4.6.3
pikepdf==2.14.2
Pillow==8.3.1
pytesseract==0.3.8
psycopg2-binary